Split up documentation in several files
The README became too large to grok at-once and wasn't a good reference
to specific components of the library, because it was all one blob of
text. By splitting it up in several files, with descriptive links to the
varying parts of documentation, it is easier to find a specific part of
documentation as reference.