Add a line ending transformer
The added line ending transformer can add any configured sequence to the
end of the lines that are passed to it.
By default, it will use the constant PHP_EOL as line ending. The value
of this context is dependent on the host operating system.

Custom sequences can be used as line endings. E.g.: 🐘

This will change a list of ['foo', 'bar', 'baz'] into ['foo🐘', 'bar🐘',
'baz🐘'].